pyproject.toml
[project]
name = "tradify"
version = "0.1.0"
description = ""
authors = [
    {name = "rwinkah",email = "rwinosa@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all] (>=0.115.12,<0.116.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)"
]

[tool.poetry]
packages = [{include = "tradify", from = "src"}]


[tool.poetry.scripts]
start-app = "tradify.server:start"

[tool.poetry.group.dev.dependencies]
uvicorn = {extras = ["standard"], version = "^0.34.3"}

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
